refactor(test): extract shared test helpers to conftest files
- Extract make_watcher()/build_watcher_hassette()/_DummyService/_TempService
  to tests/unit/core/conftest.py (was duplicated between
  test_service_watcher_coverage.py and test_service_watcher_exhausted.py,
  copies had already drifted)
- Extract build_hassette() to tests/unit/resources/conftest.py (was
  duplicated between test_service_edge_cases.py and
  test_auto_wait_dependencies_extra.py)
- Extract zdt()/TZ to tests/unit/scheduler/conftest.py (was duplicated
  across test_cron_trigger_internals.py and test_triggers_validation.py)
- Replace opaque MRO patching (type(ws).__bases__[0]) with named Service
  class in test_websocket_service_coverage.py
