+++ tests/unit/core/conftest.py
@@ -13,8 +13,11 @@ from hassette.core.app_lifecycle_service import AppLifecycleService
 from hassette.core.bus_service import BusService
 from hassette.core.command_executor import CommandExecutor
 from hassette.core.event_filter import EventFilter
+from hassette.core.service_watcher import ServiceWatcher
+from hassette.resources.restart import RestartSpec
+from hassette.resources.service import Service
 from hassette.test_utils.mock_hassette import make_mock_hassette
-from hassette.types.enums import ResourceStatus
+from hassette.types.enums import ResourceStatus, RestartType
 
 # Minimal DDL for log_records tests — intentionally omits many real columns.
 # See test_database_service_migrations.py for the canonical schema contract.
@@ -241,6 +244,71 @@ def make_executor(*, error_handler_timeout: float = 5.0) -> CommandExecutor:
     executor.task_bucket = task_bucket
     executor._spawned_tasks = spawned_tasks
     return executor
+
+
+class _DummyService(Service):
+    """Minimal concrete Service for watcher-level tests."""
+
+    restart_spec: RestartSpec = RestartSpec(restart_type=RestartType.TRANSIENT)
+
+    async def serve(self) -> None:
+        try:
+            await asyncio.Event().wait()
+        except asyncio.CancelledError:
+            raise
+
+
+class _TempService(Service):
+    """TEMPORARY restart type service for EXHAUSTED_DEAD tests."""
+
+    restart_spec: RestartSpec = RestartSpec(restart_type=RestartType.TEMPORARY)
+
+    async def serve(self) -> None:
+        try:
+            await asyncio.Event().wait()
+        except asyncio.CancelledError:
+            raise
+
+
+def build_watcher_hassette(*, strict_lifecycle: bool = False) -> AsyncMock:
+    """Minimal Hassette stub for ServiceWatcher unit tests."""
+    hassette = make_mock_hassette(
+        sealed=False,
+        strict_lifecycle=strict_lifecycle,
+        lifecycle={"resource_shutdown_timeout_seconds": 1, "task_cancellation_timeout_seconds": 1},
+    )
+    hassette.send_event = AsyncMock()
+    hassette.shutdown = AsyncMock()
+    return hassette
+
+
+def make_watcher(hassette: MagicMock) -> ServiceWatcher:
+    """Build a ServiceWatcher bypassing __init__ (no real Bus child needed)."""
+    watcher = ServiceWatcher.__new__(ServiceWatcher)
+    watcher.ready_event = asyncio.Event()
+    watcher.shutdown_event = asyncio.Event()
+    watcher._ready_reason = None
+    watcher._status = ResourceStatus.NOT_STARTED
+    watcher._previous_status = ResourceStatus.NOT_STARTED
+    watcher.shutdown_completed = False
+    watcher.shutting_down = False
+    watcher.initializing = False
+    watcher._init_task = None
+    watcher._cache = None
+    watcher.hassette = hassette
+    watcher.parent = hassette
+    watcher.children = []
+    watcher._budgets = {}
+    watcher._restarting = set()
+    watcher._cooldown_tasks = {}
+    watcher._cooldown_cycles = {}
+    watcher.logger = logging.getLogger("hassette.test.service_watcher")
+    task_bucket = MagicMock()
+    task_bucket.spawn = Mock(side_effect=lambda coro, **_kw: asyncio.create_task(coro))
+    watcher.task_bucket = task_bucket
+    watcher.bus = MagicMock()
+    watcher.bus.on = AsyncMock()
+    return watcher
 
 
 def make_bus_service(*, config_timeout: float | None = 600.0, max_concurrent_dispatches: int = 50) -> BusService:
