+++ tests/unit/core/test_service_watcher_coverage.py
@@ -7,75 +7,20 @@ listener registration, the BusService-recovery gate, on_service_running's early-
 guards, cooldown abort/failure paths, and the multiple-services-found warning path.
 """
 
-import asyncio
-import logging
-from typing import ClassVar
 from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
 import pytest
 
 from hassette.core.bus_service import BusService
-from hassette.core.service_watcher import ServiceWatcher
 from hassette.events import HassetteServiceEvent
 from hassette.events.base import HassettePayload
 from hassette.events.hassette import ServiceStatusPayload
 from hassette.resources.restart import RestartSpec
-from hassette.resources.service import Service
 from hassette.test_utils import make_mock_hassette, make_service_failed_event, make_service_running_event
 from hassette.types import ResourceStatus, Topic
 from hassette.types.enums import RestartType
 
-
-def build_hassette() -> AsyncMock:
-    """Minimal Hassette stub for ServiceWatcher unit tests (mirrors test_service_watcher_exhausted.py)."""
-    hassette = make_mock_hassette(
-        sealed=False,
-        lifecycle={"resource_shutdown_timeout_seconds": 1, "task_cancellation_timeout_seconds": 1},
-    )
-    hassette.send_event = AsyncMock()
-    hassette.shutdown = AsyncMock()
-    return hassette
-
-
-class _DummyService(Service):
-    """Minimal concrete Service for watcher-level tests."""
-
-    restart_spec: ClassVar[RestartSpec] = RestartSpec(restart_type=RestartType.TRANSIENT)
-
-    async def serve(self) -> None:
-        try:
-            await asyncio.Event().wait()
-        except asyncio.CancelledError:
-            raise
-
-
-def make_watcher(hassette: MagicMock) -> ServiceWatcher:
-    """Build a ServiceWatcher bypassing __init__ (no real Bus child needed)."""
-    watcher = ServiceWatcher.__new__(ServiceWatcher)
-    watcher.ready_event = asyncio.Event()
-    watcher.shutdown_event = asyncio.Event()
-    watcher._ready_reason = None
-    watcher._status = ResourceStatus.NOT_STARTED
-    watcher._previous_status = ResourceStatus.NOT_STARTED
-    watcher.shutdown_completed = False
-    watcher.shutting_down = False
-    watcher.initializing = False
-    watcher._init_task = None
-    watcher._cache = None
-    watcher.hassette = hassette
-    watcher.parent = hassette
-    watcher.children = []
-    watcher._budgets = {}
-    watcher._restarting = set()
-    watcher._cooldown_tasks = {}
-    watcher._cooldown_cycles = {}
-    watcher.logger = logging.getLogger("hassette.test.service_watcher")
-    task_bucket = MagicMock()
-    task_bucket.spawn = Mock(side_effect=lambda coro, **_kw: asyncio.create_task(coro))
-    watcher.task_bucket = task_bucket
-    watcher.bus = MagicMock()
-    watcher.bus.on = AsyncMock()
-    return watcher
+from .conftest import _DummyService, build_watcher_hassette, make_watcher
 
 
 class TestConfigLogLevel:
@@ -90,7 +35,7 @@ class TestConfigLogLevel:
 class TestOnInitialize:
     async def test_marks_ready_and_registers_listeners(self) -> None:
         """on_initialize() registers listeners then marks the watcher ready."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
 
         # boundary-exempt: collaborator of on_initialize
@@ -107,7 +52,7 @@ class TestOnInitialize:
 class TestRegisterInternalEventListeners:
     async def test_registers_five_listeners_with_correct_status_filters(self) -> None:
         """Registers restart/shutdown/log/running/bus-recovery handlers on the correct statuses."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
 
         await watcher.register_internal_event_listeners()
@@ -146,7 +91,7 @@ class TestLogServiceEvent:
 
     async def test_skips_logging_when_status_unchanged(self) -> None:
         """No transition (status == previous_status) logs at debug without a transition message."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         watcher.logger = Mock()
 
@@ -168,7 +113,7 @@ class TestLogServiceEvent:
 
     async def test_logs_transition_when_status_changed(self) -> None:
         """A real transition logs once at debug (the transition message)."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         watcher.logger = Mock()
 
@@ -194,7 +139,7 @@ class TestLogServiceEvent:
 class TestOnBusServiceRunning:
     async def test_ignores_non_bus_service_events(self) -> None:
         """Events for a resource other than BusService do not trigger reconciliation."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         watcher.reconcile_after_bus_recovery = AsyncMock()  # boundary-exempt: collaborator of on_bus_service_running
 
@@ -207,7 +152,7 @@ class TestOnBusServiceRunning:
 
     async def test_triggers_reconciliation_for_bus_service(self) -> None:
         """A RUNNING event for BusService itself triggers the reconciliation scan."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         watcher.reconcile_after_bus_recovery = AsyncMock()  # boundary-exempt: collaborator of on_bus_service_running
 
@@ -229,7 +174,7 @@ class TestOnBusServiceRunning:
 class TestOnServiceRunningEarlyReturns:
     async def test_returns_early_when_no_budget_and_not_restarting(self) -> None:
         """A RUNNING event for a service with no budget entry and no in-progress restart is a no-op."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         dummy = _DummyService(hassette)
         hassette.children = [dummy]
@@ -246,7 +191,7 @@ class TestOnServiceRunningEarlyReturns:
 
     async def test_returns_early_when_service_not_found(self) -> None:
         """A RUNNING event for a service no longer present in hassette.children is a no-op."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         dummy = _DummyService(hassette)
         # Budget exists (so the first guard passes) but the service itself is gone.
@@ -263,7 +208,7 @@ class TestOnServiceRunningEarlyReturns:
 class TestCooldownAndRetry:
     async def test_aborts_without_restart_when_shutdown_requested(self) -> None:
         """cooldown_and_retry does not attempt a restart if shutdown fires during the cooldown sleep."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         dummy = _DummyService(hassette)
         dummy.restart = AsyncMock()  # boundary-exempt: collaborator of cooldown_and_retry
@@ -280,7 +225,7 @@ class TestCooldownAndRetry:
 
     async def test_restart_exception_after_cooldown_is_caught(self) -> None:
         """A service.restart() failure after cooldown is logged, not propagated."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         dummy = _DummyService(hassette)
         dummy.restart = AsyncMock(side_effect=RuntimeError("restart blew up"))
@@ -296,7 +241,7 @@ class TestCooldownAndRetry:
 
     async def test_skips_restart_when_service_gone_after_cooldown(self) -> None:
         """If the service disappears during cooldown, restart is skipped without error."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         hassette.children = []
 
@@ -309,7 +254,7 @@ class TestCooldownAndRetry:
 class TestRestartServiceMultipleMatches:
     async def test_restarts_all_matching_services_and_warns(self) -> None:
         """When two services share class_name/role, restart_service restarts both."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
 
         svc_a = _DummyService(hassette)
@@ -334,7 +279,7 @@ class TestRestartServiceMultipleMatches:
 class TestRestartServiceNoServiceFound:
     async def test_returns_early_without_side_effects(self) -> None:
         """restart_service for a resource_name with no matching child is a no-op."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         hassette.children = []
 
@@ -351,7 +296,7 @@ class TestRestartServiceNoServiceFound:
 class TestShutdownIfCrashed:
     async def test_reason_omits_exception_type_when_none(self) -> None:
         """When exception_type is falsy, the fatal reason has no ': <type>' suffix."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         hassette.record_fatal_reason = Mock()
         hassette.request_shutdown = Mock()
         watcher = make_watcher(hassette)
@@ -375,7 +320,7 @@ class TestShutdownIfCrashed:
 
     async def test_reraises_on_unexpected_internal_failure(self) -> None:
         """If record_fatal_reason itself raises, shutdown_if_crashed logs and re-raises."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         hassette.record_fatal_reason = Mock(side_effect=RuntimeError("state corrupted"))
         hassette.request_shutdown = Mock()
         watcher = make_watcher(hassette)
@@ -404,7 +349,7 @@ class TestOnServiceRunningBudgetNoneBranch:
     async def test_clears_in_restart_flag_without_creating_budget(self) -> None:
         """A RUNNING event while restarting (but with no budget entry yet) clears the flag,
         without fabricating a budget."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         dummy = _DummyService(hassette)
         dummy.mark_ready(reason="test")
@@ -424,7 +369,7 @@ class TestOnServiceRunningBudgetNoneBranch:
 class TestReconcileAfterBusRecoverySkips:
     async def test_skips_non_service_children(self) -> None:
         """Non-Service children (e.g. plain resources) are ignored by the reconciliation scan."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         watcher.restart_service = AsyncMock()  # boundary-exempt: collaborator of reconcile_after_bus_recovery
 
@@ -437,7 +382,7 @@ class TestReconcileAfterBusRecoverySkips:
 
     async def test_skips_services_not_in_failed_state(self) -> None:
         """Services that are not FAILED are left alone."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         watcher.restart_service = AsyncMock()  # boundary-exempt: collaborator of reconcile_after_bus_recovery
 
@@ -451,7 +396,7 @@ class TestReconcileAfterBusRecoverySkips:
 
     async def test_skips_failed_services_with_existing_budget(self) -> None:
         """A FAILED service that already has a budget entry was handled normally — skip it."""
-        hassette = build_hassette()
+        hassette = build_watcher_hassette()
         watcher = make_watcher(hassette)
         watcher.restart_service = AsyncMock()  # boundary-exempt: collaborator of reconcile_after_bus_recovery
