+++ tests/unit/core/test_service_watcher_exhausted.py
@@ -7,54 +7,15 @@ Verifies:
 - Status assignment is skipped (with warning) when get_service returns empty list
 """
 
-import asyncio
-import logging
-from typing import ClassVar
-from unittest.mock import AsyncMock, MagicMock, Mock, patch
+from unittest.mock import MagicMock, patch
 
-from hassette.core.service_watcher import ServiceWatcher
 from hassette.events.hassette import ServiceStatusPayload
 from hassette.resources.restart import RestartSpec
 from hassette.resources.service import Service
-from hassette.test_utils import make_mock_hassette
 from hassette.types import ResourceStatus
 from hassette.types.enums import ResourceRole, RestartType
 
-
-def build_hassette(*, strict_lifecycle: bool = False) -> AsyncMock:
-    """Minimal Hassette stub for ServiceWatcher unit tests."""
-    hassette = make_mock_hassette(
-        sealed=False,
-        strict_lifecycle=strict_lifecycle,
-        lifecycle={"resource_shutdown_timeout_seconds": 1, "task_cancellation_timeout_seconds": 1},
-    )
-    hassette.send_event = AsyncMock()
-    hassette.shutdown = AsyncMock()
-    return hassette
-
-
-class _DummyService(Service):
-    """Minimal concrete Service for exhaustion tests."""
-
-    restart_spec: ClassVar[RestartSpec] = RestartSpec(restart_type=RestartType.TRANSIENT)
-
-    async def serve(self) -> None:
-        try:
-            await asyncio.Event().wait()
-        except asyncio.CancelledError:
-            raise
-
-
-class _TempService(Service):
-    """TEMPORARY restart type service for EXHAUSTED_DEAD tests."""
-
-    restart_spec: ClassVar[RestartSpec] = RestartSpec(restart_type=RestartType.TEMPORARY)
-
-    async def serve(self) -> None:
-        try:
-            await asyncio.Event().wait()
-        except asyncio.CancelledError:
-            raise
+from .conftest import _DummyService, _TempService, build_watcher_hassette, make_watcher
 
 
 def make_failed_payload(service: Service) -> ServiceStatusPayload:
@@ -72,38 +33,9 @@ def make_failed_payload(service: Service) -> ServiceStatusPayload:
     )
 
 
-def make_watcher(hassette: MagicMock) -> ServiceWatcher:
-    """Build a ServiceWatcher that skips its Bus child (not needed for unit tests)."""
-    watcher = ServiceWatcher.__new__(ServiceWatcher)
-    # Initialize LifecycleMixin state
-    watcher.ready_event = asyncio.Event()
-    watcher.shutdown_event = asyncio.Event()
-    watcher._ready_reason = None
-    watcher._status = ResourceStatus.NOT_STARTED
-    watcher._previous_status = ResourceStatus.NOT_STARTED
-    watcher.shutdown_completed = False
-    watcher.shutting_down = False
-    watcher.initializing = False
-    watcher._init_task = None
-    watcher._cache = None
-    watcher.hassette = hassette
-    watcher.parent = hassette
-    watcher.children = []
-    watcher._budgets = {}
-    watcher._restarting = set()
-    watcher._cooldown_tasks = {}
-    watcher._cooldown_cycles = {}
-    watcher.logger = logging.getLogger("hassette.test.service_watcher")
-    # Task bucket mock
-    task_bucket = MagicMock()
-    task_bucket.spawn = Mock(side_effect=lambda coro, **_kw: asyncio.create_task(coro))
-    watcher.task_bucket = task_bucket
-    return watcher
-
-
 async def test_exhausted_cooling_sets_status_on_instance():
     """TRANSIENT service: handle_exhaustion sets EXHAUSTED_COOLING on the service instance."""
-    hassette = build_hassette()
+    hassette = build_watcher_hassette()
     service = _DummyService(hassette)
     service._status = ResourceStatus.FAILED
     hassette.children = [service]
@@ -141,7 +73,7 @@ async def test_exhausted_cooling_sets_status_on_instance():
 
 async def test_exhausted_dead_sets_status_on_instance():
     """TEMPORARY service: handle_exhaustion sets EXHAUSTED_DEAD on the service instance."""
-    hassette = build_hassette()
+    hassette = build_watcher_hassette()
     service = _TempService(hassette)
     service._status = ResourceStatus.FAILED
     hassette.children = [service]
@@ -165,7 +97,7 @@ async def test_exhausted_dead_sets_status_on_instance():
 
 async def test_cooldown_exceeded_sets_exhausted_dead():
     """cooldown_and_retry transitions EXHAUSTED_COOLING → EXHAUSTED_DEAD when max_cooldown_cycles exceeded."""
-    hassette = build_hassette()
+    hassette = build_watcher_hassette()
     service = _DummyService(hassette)
     service._status = ResourceStatus.EXHAUSTED_COOLING
     hassette.children = [service]
@@ -194,7 +126,7 @@ async def test_cooldown_exceeded_sets_exhausted_dead():
 
 async def test_exhausted_status_skipped_when_service_not_found():
     """When get_service returns empty list, status set is skipped — no exception, event still emitted."""
-    hassette = build_hassette()
+    hassette = build_watcher_hassette()
     hassette.children = []
 
     watcher = make_watcher(hassette)
